[breakingchange] 디렉토리 구조 변경
auction_product와 product 테이블이 합쳐지게 되면서 global과 domain으로 나누었던 디렉토리 구조를 하나로 합치고, 불필요했던 config 파일을 삭제했다. 그리고 새로 기능들을 작성하기 위해 기존에 있던 서비스 클래스를 삭제했다.
