+++ src/main/java/f4/product/constant/AuctionStatus.java
@@ -1,0 +1,8 @@
+package f4.product.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum AuctionStatus {
+  WAIT, PROGRESS, END;
+}
