+++ src/main/java/f4/product/constant/CustomErrorCode.java
@@ -1,4 +1,4 @@
-package f4.product.global.constant;
+package f4.product.constant;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
