--- src/main/java/f4/product/domain/auction/persist/repository/AuctionRepository.java
@@ -1,10 +1,0 @@
-package f4.product.domain.auction.persist.repository;
-
-import f4.product.domain.auction.persist.entity.Auction;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuctionRepository extends JpaRepository<Auction, Long> {
-
-}
