--- src/main/java/f4/product/domain/product/config/aws/exception/EmptyFileException.java
@@ -1,11 +1,0 @@
-package f4.product.domain.product.config.aws.exception;
-
-import f4.product.global.constant.CustomErrorCode;
-import f4.product.global.exception.CustomException;
-
-public class EmptyFileException extends CustomException {
-
-  public EmptyFileException(CustomErrorCode customErrorCode) {
-    super(customErrorCode);
-  }
-}
