--- src/main/java/f4/product/domain/product/constant/AuctionStatus.java
@@ -1,8 +1,0 @@
-package f4.product.domain.product.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum AuctionStatus {
-  PROGRESS, WAIT, COMPLETE;
-}
