--- src/main/java/f4/product/domain/product/constant/SellStatus.java
@@ -1,8 +1,0 @@
-package f4.product.domain.product.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum SellStatus {
-  SELL, SOLD;
-}
