--- src/main/java/f4/product/global/config/MapperConfig.java
@@ -1,13 +1,0 @@
-package f4.product.global.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MapperConfig {
-  @Bean
-  public ModelMapper modelMapper() {
-    return new ModelMapper();
-  }
-}
