+++ src/main/java/f4/product/persist/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package f4.product.domain.product.persist.repository;
+package f4.product.persist.repository;
 
-import f4.product.domain.product.persist.entity.Product;
+import f4.product.persist.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
